Move the Analysis Server Example
The image deploy tool currently picks up everything matching sintr_* which was also collecting this example. This cluttered up the worker nodes
