+++ analysis_server_example/bin/decompress.dart
