+++ analysis_server_example/bin/query_wait_for_done.dart
