+++ analysis_server_example/lib/stacktrace_model.dart
